User profile verification

+++ src/api-v1/users/profile/profile.controller.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 import { PrismaClient } from "@prisma/client";
 import { validateUpdateUserBody } from "../../../middleware/validator";
+let jwt = require("jsonwebtoken");
 
 const prisma = new PrismaClient();
 
@@ -38,29 +39,42 @@ export default class ProfileController {
   public deleteUser =async (req:Request, res: Response) => {
       try{
         const {user_id} = req.body;
-        const checkUser = await prisma.user.findFirst({
-            where:{
-                user_id,
+            let token = req.headers["x-access-token"] || req.headers["authorization"];
+            token = token.slice(7, token.length);
+
+        jwt.verify(token, process.env.SECRET, async (err: any, decoded: any) =>{
+            if(decoded.userId == user_id){
+                const checkUser = await prisma.user.findFirst({
+                    where:{
+                        user_id,
+                    }
+                });
+                if(checkUser){
+                    const user = await prisma.user.delete({
+                        where:{
+                            user_id,
+                        }
+                    });
+                    res.status(200).send({
+                        deleteUserResponse: true,
+                        user,
+                    });
+                }
+                else{
+                    res.status(400).send({
+                        deleteUserResponse: false,
+                        errorMessage: "Invalid User ID",
+                        
+                    })
+                }
+            }
+            else{
+                res.status(403).send({
+                    errorMessage: "Invalid access",
+                })
             }
         });
-        if(checkUser){
-            const user = await prisma.user.delete({
-                where:{
-                    user_id,
-                }
-            });
-            res.status(200).send({
-                deleteUserResponse: true,
-                user,
-            });
-        }
-        else{
-            res.status(400).send({
-                deleteUserResponse: false,
-                errorMessage: "Invalid User ID",
-                
-            })
-        }
+
         
       }catch(e){
         console.error(e);
@@ -74,39 +88,53 @@ export default class ProfileController {
   public updateUserDets =async (req:Request, res:Response) => {
       try{
         const {user_id} = req.body;
-        const checkUser = await prisma.user.findFirst({
-            where:{
-                user_id,
-            }
-        });
-        if(checkUser){
-            const {name, avatar, institute, department, year} = req.body;
-            console.log(name);
-            if(validateUpdateUserBody(req, res)){
-                const user = await prisma.user.update({
+        let token = req.headers["x-access-token"] || req.headers["authorization"];
+        token = token.slice(7, token.length);
+
+        jwt.verify(token, process.env.SECRET, async (err: any, decoded: any) =>{
+
+            if(decoded.userId == user_id){
+                const checkUser = await prisma.user.findFirst({
                     where:{
                         user_id,
-                    },
-                    data:{
-                        name,
-                        avatar,
-                        institute,
-                        department,
-                        year,
                     }
-                })
-                delete user['password'];
-                res.status(200).send({
-                    updateUserResponse: true,
-                    user,
-                })
+                });
+                if(checkUser){
+                    const {name, avatar, institute, department, year} = req.body;
+                    console.log(name);
+                    if(validateUpdateUserBody(req, res)){
+                        const user = await prisma.user.update({
+                            where:{
+                                user_id,
+                            },
+                            data:{
+                                name,
+                                avatar,
+                                institute,
+                                department,
+                                year,
+                            }
+                        })
+                        delete user['password'];
+                        res.status(200).send({
+                            updateUserResponse: true,
+                            user,
+                        })
+                    }
+                }else{
+                    res.status(400).send({
+                        updateUserResponse: false,
+                        errorMessage: "Invalid user ID",
+                    })
+                }
             }
-        }else{
-            res.status(400).send({
-                updateUserResponse: false,
-                errorMessage: "Invalid user ID",
-            })
-        }
+            else{
+                res.status(403).send({
+                    errorMessage: "Invalid access",
+                })          
+            }
+        });
+
       }catch(e){
         console.error(e);
         res.status(500).send({
